Modularizacion de rutas
